requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
gunicorn
psycopg2-binary
mysql-connector-python